ARROW-10801: [Rust] [Flight] Support sending FlightData for Dictionaries with that of a RecordBatch
This PR contains a series of refactorings to extract code from the IPC `FileWriter` and `StreamWriter` that generates `EncodedData` instances. I highly recommend reviewing commit-by-commit, I tried to make the changes fairly mechanical and easy to understand at each step.

With those refactorings, it becomes easier to reuse the code with Flight too, so that when we're generating `FlightData` for a `RecordBatch`, we get the `FlightData` for the batch's dictionaries too.

This will help in getting the Flight integration tests that include dictionaries to pass... I'm also having some unrelated protocol problems over in [this PR I'm working on](https://github.com/apache/arrow/pull/8641) but this work will help there.

Closes #8826 from carols10cents/refactor-file-stream-writer

Authored-by: Carol (Nichols || Goulding) <carol.nichols@gmail.com>
Signed-off-by: Neville Dipale <nevilledips@gmail.com>