ARROW-11194: [Rust] Enable packed_simd for aarch64
packed_simd has support for aarch64 for the functions that we are using.

We can't test this feature yet, as we don't have aarch64 targets on our Rust CI yet. I however tested this on an ARM Mac device.
I perceive the build risk to be low, as I'm only enabling `aarch64` in addition to the already used `x86_64` target in `packed_simd`.

Closes #9148 from nevi-me/ARROW-11194

Authored-by: Neville Dipale <nevilledips@gmail.com>
Signed-off-by: Paddy Horan <paddyhoran@hotmail.com>