ARROW-11376: [C++] ThreadedTaskGroup failure with Thread Sanitizer enabled
In the test we were checking if the outer task set some varable before waiting on the task group.  This was a potential data race and potential false positive.  I fixed the initial condition to false to avoid false positives and then moved the check after the finish to avoid the data race.  Also, while investigating, I noticed that the ok() method could be const but wasn't (at one point I thought this might be related to the issue) so I changed that.

Closes #9319 from westonpace/bugfix/arrow-11371

Authored-by: Weston Pace <weston.pace@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>