ARROW-11394: [Rust] Tests for Slice & Concat
I took a stab at adding some tests for concat (and the underlying MutableArrayData) that cover the use case of the array having an offset. I had some problems with the assertions on the string array which makes me think this isn't fully fixed, but I wanted to put it up and see if anyone had suggestions for how to address this.

Closes #9339 from bjchambers/ARROW-11394-slice-concat

Authored-by: Ben Chambers <bchambers@apache.org>
Signed-off-by: Jorge C. Leitao <jorgecarleitao@gmail.com>