ARROW-11908: [C++] Work around flaky integration by retrying
In principle I don't like tests which retry until they succeed, but in the meantime, flaky integration tests are making them a bad signal for maintainers. While the Rust side gets sorted out, this should paper over the issues. (It doesn't help that 'connected' in gRPC is (for reasonable reasons) nebulous, such that you don't get an actual connection error until you make a call.)

(The main issue is that it appears Rust's gRPC server returns to application code before it's quite fully listening, so a fast-enough client like C++ might run into a server that's not quite ready yet.)

Closes #9755 from lidavidm/arrow-11908

Authored-by: David Li <li.davidm96@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>