ARROW-11887: [C++] Add asynchronous read to streaming CSV reader
This moves the read to the IO context and runs continuations on the CPU thread pool.  It does not add any parallelism.  The resulting reader is not reentrant or async-reentrant and it does not do any fan-out parallelism for converting/etc.  That follow-up can be addressed in ARROW-11889.

This is needed by ARROW-7001 to enable nested parallelism in the datasets API.

Closes #9644 from westonpace/feature/arrow-11887

Authored-by: Weston Pace <weston.pace@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>